added access_ops.py util
